+++ argilla-server/src/argilla_server/enums.py
@@ -81,8 +81,11 @@ class MetadataPropertyType(str, Enum):
 
 
 class RecordSortField(str, Enum):
+    id = "id"
+    external_id = "external_id"
     inserted_at = "inserted_at"
     updated_at = "updated_at"
+    status = "status"
 
 
 class SortOrder(str, Enum):